ToDoList/requirements.txt
django
djangorestframework
gunicorn
django-cors-headers
pytest-django